+++ apps/vscode-wing/src/constants.ts
@@ -1,0 +1,7 @@
+export const LANGUAGE_SERVER_NAME = "Wing Language Server";
+export const LANGUAGE_SERVER_ID = "wing-language-server";
+export const COMMAND_OPEN_CONSOLE = "wing.openConsole";
+export const COMMAND_OPEN_FILE = "wing.openFile";
+export const CFG_WING = "wing";
+export const CFG_WING_BIN = `${CFG_WING}.bin`;
+export const VIEW_TYPE_CONSOLE = "wing.console";
